Module futures01::executor

Executors

This module contains tools for managing the raw execution of futures, which is needed when building executors (places where futures can run).

More information about executors can be found online at tokio.rs.

Structs

NotifyHandle
A NotifyHandle is the core value through which notifications are routed in the futures crate.

Traits

Notify
A trait which represents a sink of notifications that a future is ready to make progress.

UnsafeNotify
An unsafe trait for implementing custom forms of memory management behind a Task.

Functions

with_notify
Sets the NotifyHandle of the current task for the duration of the provided closure.